requirements.txt
blinker==1.8.2
click==8.1.7
Flask==3.0.3
Flask-Cors==4.0.1
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==2.1.5
peewee==3.17.6
PyMySQL==1.1.1
PyYAML==6.0.1
Werkzeug==3.0.3
